refactor: refactor contacts page and slice
- In contacts slice removed permanent tokent
- Moved redux root state and dispatch in separate file
- Fixed imports in other slices
- Refactored save handler on the contacts page
