+++ src/hooks/redux-hooks.ts
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector, TypedUseSelectorHook } from 'react-redux';
-import { RootState } from '../store';
-import { AppDispatch } from '../store';
+import { RootState } from '../store/types';
+import { AppDispatch } from '../store/types';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
