+++ src/store/index.ts
@@ -21,7 +21,4 @@ const store = configureStore({
     }
 });
 
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
-
 export default store;
