+++ src/store/types.ts
@@ -1,0 +1,4 @@
+import store from './index';
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
